Fix static_cast that triggers compile errors with -Wpedantic (#1632)
(cherry picked from commit 645dd5485b2c35ea2a5a7b3433c838585e6c5972)

# Conflicts:
#	src/Camera.cc
#	src/ParticleEmitter.cc
#	src/Sensor.cc

+++ src/ParticleEmitter.cc
@@ -264,7 +264,7 @@ bool ParticleEmitter::SetType(const std::string &_typeStr)
 /////////////////////////////////////////////////
 std::string ParticleEmitter::TypeStr() const
 {
-  size_t index = static_cast<int>(this->dataPtr->type);
+  size_t index = static_cast<size_t>(this->dataPtr->type);
   if (index < emitterTypeStrs.size())
     return emitterTypeStrs[index];
   return "point";
